% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box2.R
\name{Box2UI}
\alias{Box2UI}
\title{Box2 UI module}
\usage{
Box2UI(id, run.box.title = "title", run.box.color = "warning",
  button.label = "Calculate", button.icon = shiny::icon("refresh"),
  button.style = "unite", button.color = "primary",
  button.size = "sm", description.box.title = "Description",
  description.box.color = "primary", path.html.document = NULL)
}
\arguments{
\item{id}{Shiny id}

\item{run.box.title}{header title of the run box}

\item{run.box.color}{run box outline color}

\item{button.label}{run button label}

\item{button.icon}{run button icon}

\item{button.style}{run button style refer \code{\link[shinyWidgets]{actionBttn}}}

\item{button.color}{run button color refer \code{\link[shinyWidgets]{actionBttn}}}

\item{button.size}{run button size refer \code{\link[shinyWidgets]{actionBttn}}}

\item{description.box.title}{description box header title}

\item{description.box.color}{description box color}

\item{path.html.document}{path for html document of the description}
}
\value{
Shiny UI
}
\description{
Shiny UI Module for use with \link{Box2}
}
\details{
Call via \code{Box2UI("your_id")}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
