% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QBox.R
\name{QBox}
\alias{QBox}
\title{QBox server module}
\usage{
QBox(input, output, session, submit.button.label = "Submit Again")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{submit.button.label}{The submit button title}
}
\value{
Shiny UI
}
\description{
Shiny server Module for qualitative page
}
\details{
Call via \code{shiny::callModule(QBox, "your_id", ...)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
