% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QBox.R
\name{QBoxUI}
\alias{QBoxUI}
\title{QBox UI module}
\usage{
QBoxUI(id, question.box.1.title = "Input Data",
  question.box.1.status = "primary",
  question.box.1.tab.1 = "Input Data Validation",
  question.box.1.tab.2 = "Data Archiving",
  question.box.2.title = "Model Use",
  question.box.2.status = "success",
  question.box.2.tab.1 = "Input Data Validation",
  question.box.2.tab.2 = "Data Archiving",
  question.box.3.title = "ModelGovernance",
  question.box.3.status = "danger",
  question.box.3.tab.1 = "Input Data Validation",
  question.box.3.tab.2 = "Data Archiving",
  question.box.3.tab.3 = "Input Data Validation",
  question.box.3.tab.4 = "Data Archiving",
  question.box.4.title = "Model Design",
  question.box.4.status = "warning",
  question.box.4.tab.1 = "Input Data Validation",
  question.box.4.tab.2 = "Data Archiving",
  submit.button.label = "Submit")
}
\arguments{
\item{id}{Shiny id}

\item{question.box.1.title}{question box 1 title}

\item{question.box.1.status}{question box 1 status}

\item{question.box.1.tab.1}{tab name}

\item{question.box.1.tab.2}{tab name}

\item{question.box.2.title}{question box 2 title}

\item{question.box.2.status}{question box 2 status}

\item{question.box.2.tab.1}{tab name}

\item{question.box.2.tab.2}{tab name}

\item{question.box.3.title}{question box 3 title}

\item{question.box.3.status}{question box 3 status}

\item{question.box.3.tab.1}{tab name}

\item{question.box.3.tab.2}{tab name}

\item{question.box.3.tab.3}{tab name}

\item{question.box.3.tab.4}{tab name}

\item{question.box.4.title}{question box 4 title}

\item{question.box.4.status}{question box 4 status}

\item{question.box.4.tab.1}{tab name}

\item{question.box.4.tab.2}{tab name}

\item{submit.button.label}{label for the submit button}
}
\value{
Shiny UI
}
\description{
Shiny UI Module for qualitative page
}
\details{
Call via \code{QBoxUI("your_id", ...)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
