% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreBox.R
\name{ScoreBox}
\alias{ScoreBox}
\title{ScoreBox UI module}
\usage{
ScoreBox(id, score.box.title = NULL, progress.bar.title = "Status",
  progress.bar.status = "primary", progress.bar.value = 0,
  knob.title = "Score", knob.value = 0, knob.color = "#8E44AD")
}
\arguments{
\item{id}{Shiny id}

\item{score.box.title}{header title of the score box}

\item{progress.bar.title}{progress bar title}

\item{progress.bar.status}{the status of the progress bar which determines the colors. Eg: "primary", "info", "warning", "danger", "success"}

\item{progress.bar.value}{the value of the progress bar}

\item{knob.title}{title of the knob}

\item{knob.value}{value of the knob}

\item{knob.color}{color of the knob}
}
\value{
Shiny UI
}
\description{
Shiny UI Module for showing scores
}
\details{
Call via \code{ScoreBox("your_id", ...)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
